Fix signatures with argument types 'long'.
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1262 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/xmlvm2objc/compat-lib/objc/xmlvm.h
@@ -43,10 +43,7 @@
 
 void xmlvm_init();
 
-//TODO JAVA_LONG should be typedef'ed to long long, but when we do that,
-// weird things happen with some applications. Before we figure out what
-// is going wrong, we leave it for long.
-typedef long JAVA_LONG;
+typedef long long JAVA_LONG;
 
 typedef union {
     id        o;
